\name{nAnomalyDetectionTs}
\alias{nAnomalyDetectionTs}
\title{
nAnomalyDetectionTs
}
\description{
Passes the arguments to the AnomalyDetectionTs function, and then returns just the anomalies
}
\usage{
nAnomalyDetectionTs(data)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{data}{
See documentation of nAnomalyDetectionTs
}
}
\details{
This is a fairly simple wrapper around AnomalyDetectionTs. It is needed because OpenCPU encourages
all functions to be in one package.
}
\value{
The list of anomalous times, as described by AnomalyDetectionTs documentation
}
\references{
}
\author{
Next Century Corporation
}
\note{

}

\seealso{
}
\examples{
}
