\name{NeonAngularDemo-package}
\alias{NeonAngularDemo-package}
\alias{NeonAngularDemo}
\docType{package}
\title{
Functions needed for the Neon Angular demo
}
\description{
OpenCPU strongly encourages all functions needed by an application to be in the same package.
}
\details{
\tabular{ll}{
Package: \tab NeonAngularDemo\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-06-26\cr
License: \tab Apache Version 2.0\cr
}
}
\author{
Next Century Corporation
}
\references{
}
\keyword{ package }
\seealso{
}
\examples{
}
