\name{nstl2}
\alias{nstl2}
\title{
nstl2
}
\description{
Takes the square root of the x vector and passes it to the stl2 function, and then returns just the data part of the result.
}
\usage{
nstl2(x, n.p, t.degree, t.window, s.window, s.degree, outer)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
See documentation of stl2
}
  \item{n.p}{
See documentation of stl2
}
  \item{t.degree}{
See documentation of stl2
}
  \item{t.window}{
See documentation of stl2
}
  \item{s.window}{
See documentation of stl2
}
  \item{s.degree}{
See documentation of stl2
}
  \item{outer}{
See documentation of stl2
}
}
\details{
This is a fairly simple wrapper around stl2. It is needed because OpenCPU makes it difficult to chain functions.
}
\value{
See documentation of stl2$data
}
\references{
}
\author{
Next Century Corporation
}
\note{

}

\seealso{
}
\examples{
}
